+++ src/Controller/CRUDController.php
@@ -31,7 +31,6 @@ use Sonata\AdminBundle\Util\AdminObjectAclData;
 use Sonata\AdminBundle\Util\AdminObjectAclManipulator;
 use Sonata\Exporter\Exporter;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Form\FormError;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\FormRenderer;
 use Symfony\Component\Form\FormView;
@@ -1341,6 +1340,9 @@ class CRUDController extends AbstractController
         return $translator->trans($id, $parameters, $domain, $locale);
     }
 
+    /**
+     * @psalm-suppress PossiblyUndefinedMethod https://github.com/psalm/psalm-plugin-symfony/pull/243
+     */
     protected function handleXmlHttpRequestErrorResponse(Request $request, FormInterface $form): ?JsonResponse
     {
         if ([] === array_intersect(['application/json', '*/*'], $request->getAcceptableContentTypes())) {
@@ -1349,7 +1351,6 @@ class CRUDController extends AbstractController
 
         $errors = [];
         foreach ($form->getErrors(true) as $error) {
-            \assert($error instanceof FormError);
             $errors[] = $error->getMessage();
         }
 
@@ -1403,7 +1404,7 @@ class CRUDController extends AbstractController
     }
 
     /**
-     * @phpstan-return array{_tab?: string}
+     * @return array{_tab?: string}
      */
     final protected function getSelectedTab(Request $request): array
     {