+++ tests/Fixtures/Entity/FooToStringNull.php
@@ -15,9 +15,9 @@ namespace Sonata\AdminBundle\Tests\Fixtures\Entity;
 
 final class FooToStringNull
 {
-    // In case __toString returns an attribute not yet set
     public function __toString()
     {
+        // In case __toString returns an attribute not yet set
         return null;
     }
 }